phase-d(bun_runtime): uws_handlers BunListener on_handshake_no_ext: convert sys verify_err to bun_uws mirror for inherent NewSocket::on_handshake

+++ src/runtime/webcore/FileReader.rs
@@ -156,7 +156,7 @@ impl Lazy {
                     {
                         if fd.stdio_tag().is_none() {
                             is_nonblocking = match sys::get_fcntl_flags(fd) {
-                                Ok(flags) => (flags & sys::O::NONBLOCK) != 0,
+                                Ok(flags) => (flags & sys::O::NONBLOCK as isize) != 0,
                                 Err(_) => false,
                             };
                         }
